Lots of boiler code to test stuff
Tons of boiler code has been added to support a majority of the
different data formats. Once a macro is written, all of this added code
will pretty much go a way. Several new tests have also been added to
test out the codes ability to read different formats. They are in no
ways complete in their tests of being able to parse data but it is
definitely a start. The float test cases need to be improved where near
machine precision is tested for both floating point cases.