Replaced code boiler plate with a macro
All of the redundant code has been replaced with a single macro that
now resides in the macro_src.rs file.

All of the parsing is now located in one location making it that much
easier to update the existing code base.

The explicit data type load_txt functions still exist though.
A simple helper function might be created to make this a bit simpler
later on as well so only one function has to be called, but the data
type is specified for it.