New parsing backend complete
The new parsing backend now works. I've refactored all of the macros, so they're much simpler now. The main logic now lays in the parse_txt function that is exposed to the user, and so the macros just have to make a call to the one function. The new parsing function also now opens up the ability to load any type that supports the FromStr trait. Although, I still need to add that generic function to the library.