Comments optional and other fixes
Comments are now optional for ReaderParams based on an issue opened. The default value is still the # character. A test was added to make sure things still work when comments are set to None. Additionally, the vector size within the macros that create the results were set to number of items + 1 just to be safe and make sure that vec doesn't internally increase its internal storage when we add in the last item.